+++ processor/src/main/java/react4j/processor/Constants.java
@@ -1,0 +1,14 @@
+package react4j.processor;
+
+final class Constants
+{
+  static final String JS_FUNCTION_CLASSNAME = "jsinterop.annotations.JsFunction";
+  static final String COMPONENT_CLASSNAME = "react4j.core.Component";
+  static final String PROCEDURE_CLASSNAME = "react4j.core.Procedure";
+  static final String EVENT_HANDLER_ANNOTATION_CLASSNAME = "react4j.annotations.EventHandler";
+  static final String REACT_COMPONENT_ANNOTATION_CLASSNAME = "react4j.annotations.ReactComponent";
+
+  private Constants()
+  {
+  }
+}