+++ processor/src/main/java/react4j/processor/package-info.java
@@ -1,4 +1,4 @@
 /**
- * This package contains the processor for the {@link react4j.annotations.ReactComponent} annotation.
+ * This package contains the processor for the react4j.annotations.ReactComponent annotation.
  */
 package react4j.processor;