+++ processor/src/test/resources/expected/com/example/arez/ComponentWithNoAutoActionEventHandler_.java
@@ -8,12 +8,12 @@ import jsinterop.annotations.JsType;
 import jsinterop.base.Js;
 import jsinterop.base.JsPropertyMap;
 import org.realityforge.arez.annotations.ArezComponent;
+import react4j.annotations.EventHandler;
 import react4j.core.BaseContext;
 import react4j.core.BaseProps;
 import react4j.core.BaseState;
 import react4j.core.ComponentConstructorFunction;
 import react4j.core.NativeAdapterComponent;
-import react4j.core.Procedure;
 import react4j.core.React;
 import react4j.core.ReactConfig;
 import react4j.core.ReactNode;
@@ -26,7 +26,7 @@ class ComponentWithNoAutoActionEventHandler_ extends ComponentWithNoAutoActionEv
   private static final ComponentConstructorFunction<BaseProps, BaseContext> TYPE = getConstructorFunction();
 
   @Nonnull
-  private final Procedure _handleFoo = create_handleFoo();
+  private final EventHandler.Procedure _handleFoo = create_handleFoo();
 
   @Nonnull
   static ReactNode _create() {
@@ -53,13 +53,13 @@ class ComponentWithNoAutoActionEventHandler_ extends ComponentWithNoAutoActionEv
   }
 
   @Nonnull
-  static Procedure _handleFoo(@Nonnull final ComponentWithNoAutoActionEventHandler component) {
+  static EventHandler.Procedure _handleFoo(@Nonnull final ComponentWithNoAutoActionEventHandler component) {
     return ((ComponentWithNoAutoActionEventHandler_) component)._handleFoo;
   }
 
   @Nonnull
-  private Procedure create_handleFoo() {
-    final Procedure handler = () -> this.handleFoo();
+  private EventHandler.Procedure create_handleFoo() {
+    final EventHandler.Procedure handler = () -> this.handleFoo();
     if( ReactConfig.enableComponentNames() ) {
       JsObject.defineProperty( Js.cast( handler ), "name", Js.cast( JsPropertyMap.of( "value", "ComponentWithNoAutoActionEventHandler.handleFoo" ) ) );
     }