+++ processor/src/test/resources/expected/com/example/event_handler/BasicHandlerComponent_.java
@@ -6,12 +6,12 @@ import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import jsinterop.base.Js;
 import jsinterop.base.JsPropertyMap;
+import react4j.annotations.EventHandler;
 import react4j.core.BaseContext;
 import react4j.core.BaseProps;
 import react4j.core.BaseState;
 import react4j.core.ComponentConstructorFunction;
 import react4j.core.NativeAdapterComponent;
-import react4j.core.Procedure;
 import react4j.core.React;
 import react4j.core.ReactConfig;
 import react4j.core.ReactNode;
@@ -21,7 +21,7 @@ class BasicHandlerComponent_ extends BasicHandlerComponent {
   private static final ComponentConstructorFunction<BaseProps, BaseContext> TYPE = getConstructorFunction();
 
   @Nonnull
-  private final Procedure _handleFoo = create_handleFoo();
+  private final EventHandler.Procedure _handleFoo = create_handleFoo();
 
   @Nonnull
   static ReactNode _create() {
@@ -48,13 +48,13 @@ class BasicHandlerComponent_ extends BasicHandlerComponent {
   }
 
   @Nonnull
-  static Procedure _handleFoo(@Nonnull final BasicHandlerComponent component) {
+  static EventHandler.Procedure _handleFoo(@Nonnull final BasicHandlerComponent component) {
     return ((BasicHandlerComponent_) component)._handleFoo;
   }
 
   @Nonnull
-  private Procedure create_handleFoo() {
-    final Procedure handler = () -> this.handleFoo();
+  private EventHandler.Procedure create_handleFoo() {
+    final EventHandler.Procedure handler = () -> this.handleFoo();
     if( ReactConfig.enableComponentNames() ) {
       JsObject.defineProperty( Js.cast( handler ), "name", Js.cast( JsPropertyMap.of( "value", "BasicHandlerComponent.handleFoo" ) ) );
     }