+++ CHANGELOG.md
@@ -8,6 +8,9 @@
   nested classes and thus a nested class could have aname collision with a generated class. Instead the
   `_` character has been used to separate name components.
 * **\[processor\]** Remove direct dependency on `elemental.core` package from `processor` artifact.
+* **\[processor\]** Remove direct dependency on `jsinterop.*` packages from `processor` artifact.
+* **\[processor\]** Remove direct dependency on `react4j.annotations` and `react4j.core` packages
+  from `processor` artifact.
 
 ### [v0.16](https://github.com/react4j/react4j/tree/v0.16) (2017-11-14)
 [Full Changelog](https://github.com/react4j/react4j/compare/v0.15...v0.16)